edit pass
Edit pass on the final article.
