+++ docs/csharp/programming-guide/classes-and-structs/snippets/object-collection-initializers/ObjectInitializersExecutionOrder.cs
