+++ docs/csharp/programming-guide/classes-and-structs/snippets/object-collection-initializers/Program.cs
@@ -1,0 +1,13 @@
+﻿namespace object_collection_initializers;
+
+class Program
+{
+    static void Main(string[] args)
+    {
+        BasicObjectInitializers.ObjectInitializerExamples();
+        HowToObjectInitializers.Main();
+        HowToIndexInitializer.Main();
+        HowToDictionaryInitializer.Main();
+        ObjectInitializersExecutionOrder.Main();
+    }
+}
