+++ docs/csharp/programming-guide/classes-and-structs/snippets/object-collection-initializers/object-collection-initializers.csproj
@@ -2,11 +2,11 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
     <RootNamespace>object_collection_initializers</RootNamespace>
     <StartupObject>object_collection_initializers.Program</StartupObject>
-    <LangVersion>11</LangVersion>
   </PropertyGroup>
 
 </Project>
