--- samples/snippets/csharp/programming-guide/classes-and-structs/object-collection-initializers/HowToClassTypedInitializer.cs
@@ -1,83 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace object_collection_initializers
-{
-    // <SnippetHowToClassTypedInitializer>
-    public class HowToClassTypedInitializer
-    {
-        public class EmbeddedClassTypeA
-        {
-            public int I { get; set; }
-            public bool B { get; set; }
-            public string S { get; set; }
-            public EmbeddedClassTypeB ClassB { get; set; }
-
-            public override string ToString() => $"{I}|{B}|{S}|||{ClassB}";
-
-            public EmbeddedClassTypeA()
-            {
-                Console.WriteLine($"Entering EmbeddedClassTypeA constructor. Values are: {this}");
-                I = 3;
-                B = true;
-                S = "abc";
-                ClassB = new() { BB = true, BI = 43 };
-                Console.WriteLine($"Exiting EmbeddedClassTypeA constructor. Values are: {this})");
-            }
-        }
-
-        public class EmbeddedClassTypeB
-        {
-            public int BI { get; set; }
-            public bool BB { get; set; }
-            public string BS { get; set; }
-
-            public override string ToString() => $"{BI}|{BB}|{BS}";
-
-            public EmbeddedClassTypeB()
-            {
-                Console.WriteLine($"Entering EmbeddedClassTypeB constructor. Values are: {this}");
-                BI = 23;
-                BB = false;
-                BS = "BBBabc";
-                Console.WriteLine($"Exiting EmbeddedClassTypeB constructor. Values are: {this})");
-            }
-        }
-
-        public static void Main()
-        {
-            var a = new EmbeddedClassTypeA
-            {
-                I = 103,
-                B = false,
-                ClassB = { BI = 100003 }
-            };
-            Console.WriteLine($"After initializing EmbeddedClassTypeA: {a}");
-
-            var a2 = new EmbeddedClassTypeA
-            {
-                I = 103,
-                B = false,
-                ClassB = new() { BI = 100003 } //New instance
-            };
-            Console.WriteLine($"After initializing EmbeddedClassTypeA a2: {a2}");
-        }
-
-        // Output:
-        //Entering EmbeddedClassTypeA constructor Values are: 0|False||||
-        //Entering EmbeddedClassTypeB constructor Values are: 0|False|
-        //Exiting EmbeddedClassTypeB constructor Values are: 23|False|BBBabc)
-        //Exiting EmbeddedClassTypeA constructor Values are: 3|True|abc|||43|True|BBBabc)
-        //After initializing EmbeddedClassTypeA: 103|False|abc|||100003|True|BBBabc
-        //Entering EmbeddedClassTypeA constructor Values are: 0|False||||
-        //Entering EmbeddedClassTypeB constructor Values are: 0|False|
-        //Exiting EmbeddedClassTypeB constructor Values are: 23|False|BBBabc)
-        //Exiting EmbeddedClassTypeA constructor Values are: 3|True|abc|||43|True|BBBabc)
-        //Entering EmbeddedClassTypeB constructor Values are: 0|False|
-        //Exiting EmbeddedClassTypeB constructor Values are: 23|False|BBBabc)
-        //After initializing EmbeddedClassTypeA a2: 103|False|abc|||100003|False|BBBabc
-    }
-    // </SnippetHowToClassTypedInitializer>
-}
