--- samples/snippets/csharp/programming-guide/classes-and-structs/object-collection-initializers/HowToIndexInitializer.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace object_collection_initializers
-{
-    // <SnippetHowToIndexInitializer>
-    public class HowToIndexInitializer
-    {
-        public class BaseballTeam
-        {
-            private string[] players = new string[9];
-            private readonly List<string> positionAbbreviations = new List<string>
-            {
-                "P", "C", "1B", "2B", "3B", "SS", "LF", "CF", "RF"
-            };
-
-            public string this[int position]
-            {
-                // Baseball positions are 1 - 9.
-                get { return players[position-1]; }
-                set { players[position-1] = value; }
-            }
-            public string this[string position]
-            {
-                get { return players[positionAbbreviations.IndexOf(position)]; }
-                set { players[positionAbbreviations.IndexOf(position)] = value; }
-            }
-        }
-
-        public static void Main()
-        {
-            var team = new BaseballTeam
-            {
-                ["RF"] = "Mookie Betts",
-                [4] = "Jose Altuve",
-                ["CF"] = "Mike Trout"
-            };
-
-            Console.WriteLine(team["2B"]);
-        }
-    }
-    // </SnippetHowToIndexInitializer>
-}
