--- samples/snippets/csharp/programming-guide/classes-and-structs/object-collection-initializers/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace object_collection_initializers
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            BasicObjectInitializers.ObjectInitializerExamples();
-            HowToObjectInitializers.Main();
-            HowToIndexInitializer.Main();
-            HowToDictionaryInitializer.Main();
-            ObjectInitializersExecutionOrder.Main();
-        }
-    }
-}
